Add standalone disruption test
This test can be run provided you have a KUBECONFIG env set and `oc` in your path somewhere.
It will run through the cluster and drain then reboot each node.
The idea here is that it will track service disruption for
services with externalTrafficPolicy Cluster,
so that we can iterate on changes without having to wait
for feedback from the entire payload blocking suites.